CR-2004-05-12-0: Bugfix:    o Repeated commlib logging when qmaster can't be                               reached                             o Fixed problem of checkin by CR-2004-05-11-1:                               PID before sge_daemonize() is not the same after                               daemonize. The /tmp/execd_messages.PID file wasn't                               appended to execd/messages file.                  Issue:     864                  Changed:   execd, gdilib                  Review:    SG
Issue number:
Obtained from:
Submitted by:
Reviewed by:

+++ source/daemons/execd/setup_execd.c
@@ -72,11 +72,9 @@ extern lList *jr_list;
 static char execd_messages_file[SGE_PATH_MAX];
 
 /*-------------------------------------------------------------------*/
-void sge_setup_sge_execd()
+void sge_setup_sge_execd(const char* tmp_err_file_name)
 {
    char err_str[1024];
-   static char tmp_err_file_name[SGE_PATH_MAX];
-   int my_pid;
 
    DENTER(TOP_LAYER, "sge_setup_sge_execd");
 
@@ -113,11 +111,13 @@ void sge_setup_sge_execd()
    sge_chdir_exit(uti_state_get_unqualified_hostname(), 1); 
    /* having passed the  previous statement we may 
       log messages into the ERR_FILE  */
-   my_pid = getpid();
-   sprintf(tmp_err_file_name,"%s."U32CFormat"",TMP_ERR_FILE_EXECD,u32c(my_pid));
-   sge_copy_append(tmp_err_file_name, ERR_FILE, SGE_MODE_APPEND);
+   if ( tmp_err_file_name != NULL) {
+      sge_copy_append((char*)tmp_err_file_name, ERR_FILE, SGE_MODE_APPEND);
+   }
    sge_switch2start_user();
-   unlink(tmp_err_file_name);
+   if ( tmp_err_file_name != NULL) {
+      unlink(tmp_err_file_name);
+   }
    sge_switch2admin_user();
    log_state_set_log_as_admin_user(1);
    sprintf(execd_messages_file, "%s/%s/%s", conf.execd_spool_dir,
